chore: retire quick-launch + community-creation UI surface (#1191)
Deletes ~5800 LOC of dead code. Quick-launch had zero course creations in
last 30 days on hf-dev; operator confirmed Communities is parked and any
future revival will go through the V6 wizard chain.

Deletes:
- /x/quick-launch page + CSS
- /api/domains/quick-launch/{,/analyze,/commit} routes
- lib/domain/quick-launch.ts (1556 LOC)
- /api/domains/suggest-name route + test (sole caller was quick-launch)
- lib/demo/content/DEMO-QUICK-001.demo.json + registry entry
- features/quick-launch.feature BDD spec
- e2e specs: communities/communities, cloud/golden-path, cloud/quick-launch,
  _archived/content-wizard, page-objects/quick-launch.page
- tests: domain/quick-launch.test.ts

Extracted to keep V5 alive:
- lib/domain/persona-loaders.ts — resolvePersonaKey + loadPersonaFlowPhases
  + loadPersonaArchetype + loadPersonaWelcomeTemplate (still used by
  course-setup.ts V5 backend + wizard-tool-executor.ts)

Repoints (8 sites):
- dashboard-config: Communities createHref → null; remove quick_launch
  RESUME_PATHS + Community quick-link
- DashboardClient: handle createHref:null guard
- jobs/page.tsx + JobsPopup.tsx: remove quick_launch JOB_TYPE_LABELS entry
- tour-definitions.ts: remove "Community" action
- smoke-test/page.tsx: remove wiz-quick item
- e2e/cloud/smoke.spec.ts: remove quick-launch test
- educator/classrooms/new: fallback "Create one" → /x/institutions/new
- task-guidance.ts: remove quick_launch guidance block
- tests/lib: drop /x/quick-launch from page-auth-coverage,
  remove "quick-launch" from tour-manifest-binding pending set,
  update course-setup.test mock to persona-loaders

Migration:
- 20260606134119_archive_quick_launch_tasks — sets UserTask.status =
  'abandoned' for any in-flight taskType='quick_launch' rows, so Jobs page
  doesn't render un-resumable orphans.

Verified: tsc error count cleanup=245 vs main=247 (no new errors introduced).

Closes #1184 (orphan-Curriculum bug becomes wontfix-by-retirement).
Closes #1191.
