+++ apps/admin/e2e/page-objects/index.ts
@@ -6,7 +6,6 @@
 export { BasePage } from './base.page';
 export { LoginPage } from './login.page';
 export { SidebarPage } from './sidebar.page';
-export { QuickLaunchPage } from './quick-launch.page';
 export { SimPage } from './sim.page';
 export { DemonstratePage } from './demonstrate.page';
 export { TeachPage } from './teach.page';
